+++ src/Resources/contao/dca/tl_form_layout.php
@@ -13,7 +13,7 @@
 $GLOBALS['TL_DCA']['tl_form_layout']['metapalettes']['bs_default extends standard'] = [];
 
 $GLOBALS['TL_DCA']['tl_form_layout']['metapalettes']['bs_horizontal extends standard'] = [
-    'bs_grid before widgets' => ['bs_label', 'bs_control', 'bs_offset']
+    'bs_grid before widgets' => ['bs_row', 'bs_label', 'bs_control', 'bs_offset']
 ];
 
 $GLOBALS['TL_DCA']['tl_form_layout']['fields']['bs_label'] = [
@@ -40,6 +40,17 @@ $GLOBALS['TL_DCA']['tl_form_layout']['fields']['bs_control'] = [
 
 $GLOBALS['TL_DCA']['tl_form_layout']['fields']['bs_offset'] = [
     'label'     => &$GLOBALS['TL_LANG']['tl_form_layout']['bs_offset'],
+    'exclude'   => true,
+    'inputType' => 'text',
+    'eval'      => array(
+        'tl_class'       => 'w50',
+        'maxlength'      => 64,
+    ),
+    'sql'       => "varchar(64) NOT NULL default ''"
+];
+
+$GLOBALS['TL_DCA']['tl_form_layout']['fields']['bs_row'] = [
+    'label'     => &$GLOBALS['TL_LANG']['tl_form_layout']['bs_row'],
     'exclude'   => true,
     'inputType' => 'text',
     'eval'      => array(